docs(no-conditional-expect): add asymmetric matchers to example usage (#847)
* docs(no-conditional-expect): add asymmetric matchers to example usage

Clarifies that conditional asymmetric matchers are considered invalid,
not just conditional assertions, as discussed in #834.

* docs(no-conditional-expect): clarify motivation for using the rule

The details should provide more than just the information that can be inferred from the name of the rule.

The incorrect usage example is also updated to be a use case where an invalid condition is caught,
better demonstrating the utility of the rule.

* update docs

---------

Co-authored-by: Verite Mugabo <mugaboverite@gmail.com>